move VirtualConsole factory and preferences implementation from core to client
- core isn't supposed to depend on client, though this is far from the only offender
